refactor(server): move and rename entities
