--- Server/Entities/CollectionRole.cs
@@ -1,9 +1,0 @@
-﻿namespace Server.Entities;
-
-public enum CollectionRole
-{
-    Viewer = 1,
-    Editor = 2,
-    Admin = 3,
-    Owner = 4
-}
