+++ Server/Entities/UserCollection.cs
@@ -18,3 +18,11 @@ public class UserCollection
 
     public CollectionRole Role { get; set; }
 }
+
+public enum CollectionRole
+{
+    Viewer = 1,
+    Editor = 2,
+    Admin = 3,
+    Owner = 4
+}
