+++ Server/Models/InviteTokenRequestDto.cs
@@ -6,14 +6,14 @@ public static class InviteTokenDto
 {
     public class Request
     {
-        public required Role AssignedRole { get; set; }
+        public required UserAppRole AssignedRole { get; set; }
     }
 
     public class Response
     {
         public required Guid Token { get; set; }
 
-        public required Role AssignedRole { get; set; }
+        public required UserAppRole AssignedRole { get; set; }
 
         public required DateTime ExpirationDate { get; set; }
     }
