+++ Server/Services/IAuthService.cs
@@ -8,5 +8,5 @@ public interface IAuthService
     Task<UserRegistrationDto.Response?> RegisterAsync(UserRegistrationDto.Request request);
     Task<RefreshTokenDto.Response?> LoginAsync(UserLoginDto.Request request);
     Task<RefreshTokenDto.Response?> RefreshTokensAsync(RefreshTokenDto.Request request);
-    Task<InviteTokenDto.Response?> CreateInviteTokenAsync(Guid createdById, Role assignedRole);
+    Task<InviteTokenDto.Response?> CreateInviteTokenAsync(Guid createdById, UserAppRole assignedRole);
 }
